requirements.txt
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.3
mpmath==1.3.0
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
sympy==1.14.0
pandas==2.2.3
